remove older logic in favor of generalized algorithm
